CLN move plot_benchmark in plotting.base
