--- benchopt/constants.py
@@ -1,8 +1,0 @@
-PLOT_KINDS = {
-    'objective_curve': 'plot_objective_curve',
-    'suboptimality_curve': 'plot_suboptimality_curve',
-    'relative_suboptimality_curve': 'plot_relative_suboptimality_curve',
-    'iteration_sub_optimality': 'plot_iteration_suboptimality_curve',
-    'iteration_curve': 'plot_iteration_curve',
-    'bar_chart': 'plot_bar_chart',
-}
